tabline: keep active tab in more usable position
When there are more tabs that can fit on the tabline, and when the
active tab is at the right end, switching to the tab to the left
of the active tab will move the active tab position to the right.

This commit changes the active tab position if the new tab is
already visible on the tab line.