Fix vim-capslock functionality after changes in CapsLockStatusline()
CapsLockStatusline() used to return '[caps]' when soft-capslock was
on... and now returns '[Caps]'.  As we're just using this to test for
the lock, we lower-case the return value in order to "normalize" the
value across versions.

The change in behaviour can be seen at:

tpope/vim-capslock@3a0f0513730f03ebabc35ab18f6d771f1d979a3f