Better fallback for transparent backgrounds
I have `hi Normal ctermbg=None` in my .vimrc to achieve a transparent background, but it turns the status line in a bloody mess (quite literally, some foreground colors become red and disgusting). This aims to address that issue.

Please note that I just recently started using Vim and that I am completely clueless when it comes to Vimscript, so this solution may not be the preferred one. Please improve it!