Airline:
Allow placement of Airline statusline in the `tabline` instead of `statusline`.  This frees up the statusline for other plugins, or potentially a vim-airline buffer manager.

The statusline is a better place for implementing buffer management, because it allows you to easily associate multiple "tabs" for a physical window much like Atom/Sublime/VSCode. This is harder to achieve with the tabline because there's only one tabline for the entire app, and vim supports horizontal splits. With this diff, the statusline can be used for buffer "tabs", which means you can visually see which tabs are open in *every* open split.