terminal: correctly shorten terminal tabline
If there is a single terminal window in  the current tabpage, the whole
long command will be displayed in the tabline, even when it should be
shortened by default.

By default the title will not be shown if there are several windows in
the current tabpage, but if there is only one single terminal window, it
will still be shown (and might be a bit long). But I guess, it is better
to just show a possible too long terminal window name, than not show
anything).

fixes #2126