Revert Section Z default value
Keep the section to commit 8929bc72a13d358bb8369443386ac3cc4796ca16,
this is, as it was before #2374 and #2385 (and how it is still described
in the readme), which has been working fine for years:

> The information in Section Z looks like this:
>
> `10% ☰ 10/100 ln : 20`

instead of reversing the symbols, adding extra characters and completely
botching spacing trying to fix the resulting rendering problems.

No reason to fix what's not broken.

Keep the added 'colnr' configuration, which was a good addition.