ci: correctly checkout vim-flavor using https://
Currently, the CI throws:

The unauthenticated git protocol on port 9418 is no longer supported.
Please see https://github.blog/2021-09-01-improving-git-protocol-security-github/ for more information.

So update vim-flavor to 4.0.1 which uses https:// links instead of
un-authenticated git protocol. Unfortunately, we have to update ruby to
version 3.0 with that.

Let's see if this works.